# Refs #33107 -- Optimized cached_import() helper.