Follow up on https://github.com/django/django/pull/14858#discussion_r709398928.

```bash
$ sudo pyperf system tune

# Before:

$ pyperf timeit --setup 'from django.utils.module_loading import cached_import' 'cached_import("django.core.cache.backends.memcached", "PyMemcacheCache")'
.....................
Mean +- std dev: 424 ns +- 16 ns

# After:

$ pyperf timeit --setup 'from django.utils.module_loading import cached_import' 'cached_import("django.core.cache.backends.memcached", "PyMemcacheCache")'
.....................
Mean +- std dev: 346 ns +- 14 ns

$ sudo pyperf system reset
```

In the worst case this was accessing `module_path` in `sys.modules` up to four times...